Fixed all error messages from StyleCop

+++ AspNetSandBox.Tests/CityCoordinatesControllerTest.cs
@@ -1,67 +1,61 @@
-using AspNetSandBox.Controllers;
-using System;
 using System.IO;
+using AspNetSandBox.Controllers;
 using Xunit;
 
 namespace AspNetSandBox.Tests
 {
+    /// <summary>Tests for getting city coordinates from city name.</summary>
     public class CityCoordinatesControllerTest
     {
+        /// <summary>Test for getting coordinates of London.</summary>
         [Fact]
         public void ConvertResponseToCityCoordinatesLondonTest()
         {
-            //Assume
-
+            // Assume
             string content = LoadJsonFromResource();
             var controller = new CityCoordinatesController();
 
-            //Act
-
+            // Act
             var output = controller.ConvertResponseToCityCoordinates(content);
 
-            //Assert
-
+            // Assert
             var cityCoordinatesForLondon = ((CityCoordinates[])output)[0];
-            Assert.Equal((decimal)-0.1257, cityCoordinatesForLondon.Longitude);
-            Assert.Equal((decimal)51.5085, cityCoordinatesForLondon.Latitude);
+            Assert.Equal(-0.1257M, cityCoordinatesForLondon.Longitude);
+            Assert.Equal(51.5085M, cityCoordinatesForLondon.Latitude);
         }
 
+        /// <summary>Test for getting coordinates of Brasov.</summary>
         [Fact]
         public void ConvertResponseToCityCoordinatesBrasovTest()
         {
-            //Assume
-
+            // Assume
             string content = LoadJsonFromResource();
             var controller = new CityCoordinatesController();
 
-            //Act
-
+            // Act
             var output = controller.ConvertResponseToCityCoordinates(content);
 
-            //Assert
-
+            // Assert
             var cityCoordinatesForBrasov = ((CityCoordinates[])output)[1];
-            Assert.Equal((decimal)25.3333, cityCoordinatesForBrasov.Longitude);
-            Assert.Equal((decimal)45.75, cityCoordinatesForBrasov.Latitude);
+            Assert.Equal(25.3333M, cityCoordinatesForBrasov.Longitude);
+            Assert.Equal(45.75M, cityCoordinatesForBrasov.Latitude);
         }
 
+        /// <summary>Test for getting coordinates of Moscow.</summary>
         [Fact]
         public void ConvertResponseToCityCoordinatesMoscowTest()
         {
-            //Assume
-
+            // Assume
             string content = LoadJsonFromResource();
             var controller = new CityCoordinatesController();
 
-            //Act
-
+            // Act
             var output = controller.ConvertResponseToCityCoordinates(content);
 
-            //Assert
-
+            // Assert
             var cityCoordinatesForMoscow = ((CityCoordinates[])output)[2];
-            Assert.Equal((decimal)37.6156, cityCoordinatesForMoscow.Longitude);
-            Assert.Equal((decimal)55.7522, cityCoordinatesForMoscow.Latitude);
+            Assert.Equal(37.6156M, cityCoordinatesForMoscow.Longitude);
+            Assert.Equal(55.7522M, cityCoordinatesForMoscow.Latitude);
         }
 
         private string LoadJsonFromResource()
